backend/requirements.txt
fastapi
pydantic
confluent-kafka
cloudinary 
python-dotenv
sqlalchemy
pymysql 
uvicorn
python-multipart
python-jose
passlib[bcrypt]
bcrypt
fastapi-mail
pydantic-settings
googlemaps
python-decouple
argon2_cffi